requirements.txt
SQLAlchemy~=2.0.36
imapclient
python-dotenv~=1.0.1
psycopg2-binary
aioimaplib~=1.1.0
fastapi~=0.115.4
uvicorn~=0.32.0
